Set the Queue-ID as early as Sendmail is willing to give it to us.
Also, pass the Queue-ID as an argument to filter_relay and filter_helo,
and document this new argument $qid.  Copy this variable to the global
$QueueID and $MsgID variables.  ($MsgID is used by md_syslog.)

We want the Queue-ID as early as possible so that we can correlate
MIMEDefang logs with the Sendmail logs that correspond to the same
message.